@@ -1,31 +1,39 @@
 {
   "name": "@superhuman/react-list",
-  "version": "0.8.6",
+  "version": "0.8.15",
   "author": "Casey Foster <c@sey.me>",
   "license": "MIT",
   "main": "react-list.js",
   "repository": {
     "type": "git",
-    "url": "https://github.com/orgsync/react-list"
+    "url": "https://github.com/caseywebdev/react-list"
   },
   "dependencies": {
     "prop-types": "15"
   },
   "peerDependencies": {
-    "react": "0.14 || 15",
-    "react-dom": "0.14 || 15"
+    "react": "0.14 || 15 || 16"
   },
   "devDependencies": {
-    "babel-eslint": "^7.1.1",
-    "babel-plugin-transform-es2015-modules-umd": "^6.12.0",
-    "babel-preset-es2015": "^6.14.0",
-    "babel-preset-react": "^6.11.1",
-    "babel-preset-stage-1": "^6.5.0",
-    "cogs": "^3.1.5",
-    "cogs-transformer-babel": "^3.0.0",
-    "cogs-transformer-eslint": "^3.0.0",
-    "cogs-transformer-replace": "^3.0.0",
-    "eslint": "^3.4.0",
-    "eslint-plugin-react": "^6.2.0"
+    "@babel/plugin-transform-modules-umd": "7",
+    "@babel/preset-env": "7",
+    "@babel/preset-react": "7",
+    "@babel/plugin-proposal-class-properties": "7",
+    "cogs-transformer-babel": "4",
+    "cogs-transformer-concat-commonjs": "3",
+    "cogs-transformer-eslint": "3",
+    "cogs-transformer-replace": "3",
+    "cogs": "4",
+    "eslint-config-coderiety": "2",
+    "react-dom": "16",
+    "react": "16"
+  },
+  "eslintConfig": {
+    "extends": "coderiety",
+    "env": {
+      "es6": true,
+      "browser": true,
+      "node": true
+    }
   }
 }
